+++ kernel/CMakeLists.txt
@@ -7,7 +7,10 @@ add_executable(kernel
         src/arch/x86/cpu/cpuid.c
         src/arch/x86/cpu/msr.c
         src/arch/x86/cpu/idt.c
+        src/arch/x86/drivers/keyboard/keyboard.c
+        src/arch/x86/drivers/timer/timer.c
         src/arch/x86/io/io.c
+        src/arch/x86/klib/klib.c
         src/arch/x86/serial/serial.c
         src/arch/x86/vga/vga.c
         src/kernel_main.c
