+++ kernel/src/arch/x86/drivers/keyboard/keyboard.h
@@ -1,0 +1,4 @@
+#pragma once
+
+// Keyboard Handler
+void keyboard_handler(void);
