+++ kernel/src/arch/x86/drivers/timer/timer.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <stdint.h>
+
+// Returns the timer
+uint32_t get_timer_ticks(void);
+
+// Timer handler
+void timer_handler(void);
