+++ kernel/src/arch/x86/klib/klib.h
@@ -1,0 +1,7 @@
+#pragma once
+
+// Sleep for given ticks
+void ksleep(int ticks);
+
+// Trap gate
+void trap(void);
